Divide by the given number of trading days
This minor change should make it possible to use the indicator on assets that are traded 24/7 or 24/3 for that matter.
However, it should be tested.
Some changes might be needed for assets that are not traded 24-hours per day.

As I only have MT5 with 24/5 assets, I'm unable to verify if this works as intended. Please test the changes on a 24/7 broker.